refactor(webpack): move utilities around
