+++ packages/webpack/src/runtime/assemble.js
